+++ include/Menu.hh
@@ -2,11 +2,13 @@
 #ifndef MENU_H_
 #define MENU_H_
 
+#include "ModulesManager.hh"
 #include "WidgetBase.hh"
 
 class Menu : public WidgetBase {
-  private:
   public:
+    ModulesManager* manager;
+
     Menu();
 
     ~Menu();