+++ modules/MapEditor/MapConfig.hh
@@ -5,11 +5,11 @@
 class MapConfig {
   private:
   public:
-    static bool show_overlay_bar;
+    static bool show_overlay_window;
     static bool show_control_window;
     static bool show_display_window;
     static bool show_graph_window;
-    static bool show_editor_window;
+    static bool show_mapeditor_window;
     static bool show_node_window;
     static bool show_inspector_window;
     MapConfig();