+++ src/BiuLab.cc
@@ -1,13 +1,16 @@
 #include "BiuLab.hh"
 
+bool   module_changed       = false;
+bool   module_selected      = false;
+string module_selected_name = "";
+
 bool               show_dock_sapce       = true;
 bool               show_main_window      = true;
 bool               show_demo_window      = false;
 bool               show_main_menu_bar    = true;
-bool               show_overlay_bar      = true;
+bool               show_overlay_window   = true;
 bool               show_control_window   = true;
 bool               show_graph_window     = true;
-bool               show_editor_window    = true;
 bool               show_node_window      = false;
 bool               show_inspector_window = true;
 bool               show_mapeditor_window = true;