+++ changelog.md
@@ -26,7 +26,8 @@
 - [ ] controller
 
 # CHANGELOG
-- [x] [v0.9] Module 2020.02.08: compete simplemap module.
+- [x] [v1.0] Core   2020.02.10: succeed in biulab v1.0.
+- [x] [v0.9] Module 2020.02.09: compete simplemap module.
 - [x] [v0.7] Module 2020.02.08: add gem elem for simple map module.
 - [x] [v0.7] Module 2020.02.08: add simple module new widget.
 - [x] [v0.7] Module 2020.02.08: fixed random function crash bug.